fix(logger): allow configuration for log level and provide your own logger
