+++ examples/esm/package.json
@@ -3,5 +3,8 @@
   "dependencies": {
     "@ar.io/sdk": "^1.1.1",
     "arweave": "^1.15.1"
+  },
+  "scripts": {
+    "postinstall": "yarn link @ar.io/sdk"
   }
 }
