+++ src/common/io.ts
@@ -49,7 +49,6 @@ import {
 } from '../types.js';
 import { AOProcess } from './contracts/ao-process.js';
 import { InvalidContractConfigurationError } from './error.js';
-import { DefaultLogger } from './logger.js';
 
 export class IO {
   static init(): AoIORead;
@@ -100,9 +99,6 @@ export class IOReadable implements AoIORead {
     } else if (isProcessIdConfiguration(config)) {
       this.process = new AOProcess({
         processId: config.processId,
-        logger: new DefaultLogger({
-          level: 'info',
-        }),
       });
     } else {
       throw new InvalidContractConfigurationError();
